% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_operations.R
\name{BusinessOperations}
\alias{BusinessOperations}
\title{Business Operations Class}
\description{
R6 class for business management operations
}
\section{Super class}{
\code{\link[SBPoptimizer:DBOperations]{SBPoptimizer::DBOperations}} -> \code{BusinessOperations}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BusinessOperations-create_business}{\code{BusinessOperations$create_business()}}
\item \href{#method-BusinessOperations-get_businesses_by_owner}{\code{BusinessOperations$get_businesses_by_owner()}}
\item \href{#method-BusinessOperations-get_business_summary}{\code{BusinessOperations$get_business_summary()}}
\item \href{#method-BusinessOperations-search_businesses}{\code{BusinessOperations$search_businesses()}}
\item \href{#method-BusinessOperations-get_business_statistics}{\code{BusinessOperations$get_business_statistics()}}
\item \href{#method-BusinessOperations-clone}{\code{BusinessOperations$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SBPoptimizer" data-topic="DBOperations" data-id="delete_record"><a href='../../SBPoptimizer/html/DBOperations.html#method-DBOperations-delete_record'><code>SBPoptimizer::DBOperations$delete_record()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SBPoptimizer" data-topic="DBOperations" data-id="get_table"><a href='../../SBPoptimizer/html/DBOperations.html#method-DBOperations-get_table'><code>SBPoptimizer::DBOperations$get_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SBPoptimizer" data-topic="DBOperations" data-id="initialize"><a href='../../SBPoptimizer/html/DBOperations.html#method-DBOperations-initialize'><code>SBPoptimizer::DBOperations$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SBPoptimizer" data-topic="DBOperations" data-id="insert_record"><a href='../../SBPoptimizer/html/DBOperations.html#method-DBOperations-insert_record'><code>SBPoptimizer::DBOperations$insert_record()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SBPoptimizer" data-topic="DBOperations" data-id="update_record"><a href='../../SBPoptimizer/html/DBOperations.html#method-DBOperations-update_record'><code>SBPoptimizer::DBOperations$update_record()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BusinessOperations-create_business"></a>}}
\if{latex}{\out{\hypertarget{method-BusinessOperations-create_business}{}}}
\subsection{Method \code{create_business()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BusinessOperations$create_business(
  owner_id,
  kra_pin,
  name,
  category,
  size,
  location,
  latitude = NULL,
  longitude = NULL,
  ward
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{owner_id}}{User ID of business owner}

\item{\code{kra_pin}}{KRA PIN number}

\item{\code{name}}{Business name}

\item{\code{category}}{Business category}

\item{\code{size}}{Business size (Small, Medium, Large)}

\item{\code{location}}{Business location}

\item{\code{latitude}}{GPS latitude (optional)}

\item{\code{longitude}}{GPS longitude (optional)}

\item{\code{ward}}{Administrative ward
Get businesses by owner}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BusinessOperations-get_businesses_by_owner"></a>}}
\if{latex}{\out{\hypertarget{method-BusinessOperations-get_businesses_by_owner}{}}}
\subsection{Method \code{get_businesses_by_owner()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BusinessOperations$get_businesses_by_owner(owner_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{owner_id}}{User ID of owner}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame of businesses
Get business summary with permit information
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BusinessOperations-get_business_summary"></a>}}
\if{latex}{\out{\hypertarget{method-BusinessOperations-get_business_summary}{}}}
\subsection{Method \code{get_business_summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BusinessOperations$get_business_summary(business_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{business_id}}{Business ID (optional, if NULL returns all)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame with business and permit summary
Search businesses
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BusinessOperations-search_businesses"></a>}}
\if{latex}{\out{\hypertarget{method-BusinessOperations-search_businesses}{}}}
\subsection{Method \code{search_businesses()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BusinessOperations$search_businesses(
  search_term = NULL,
  ward = NULL,
  category = NULL,
  size = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{search_term}}{Search term for name or KRA PIN (optional)}

\item{\code{ward}}{Ward filter (optional)}

\item{\code{category}}{Category filter (optional)}

\item{\code{size}}{Size filter (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame of matching businesses
Get business statistics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BusinessOperations-get_business_statistics"></a>}}
\if{latex}{\out{\hypertarget{method-BusinessOperations-get_business_statistics}{}}}
\subsection{Method \code{get_business_statistics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BusinessOperations$get_business_statistics()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Data frame with business statistics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BusinessOperations-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BusinessOperations-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BusinessOperations$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
