% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_operations.R
\name{ComplianceOperations}
\alias{ComplianceOperations}
\title{Compliance Operations Class}
\description{
R6 class for compliance scoring operations
}
\section{Super class}{
\code{\link[SBPoptimizer:DBOperations]{SBPoptimizer::DBOperations}} -> \code{ComplianceOperations}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ComplianceOperations-update_compliance_score}{\code{ComplianceOperations$update_compliance_score()}}
\item \href{#method-ComplianceOperations-get_compliance_scores}{\code{ComplianceOperations$get_compliance_scores()}}
\item \href{#method-ComplianceOperations-get_high_risk_businesses}{\code{ComplianceOperations$get_high_risk_businesses()}}
\item \href{#method-ComplianceOperations-clone}{\code{ComplianceOperations$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SBPoptimizer" data-topic="DBOperations" data-id="delete_record"><a href='../../SBPoptimizer/html/DBOperations.html#method-DBOperations-delete_record'><code>SBPoptimizer::DBOperations$delete_record()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SBPoptimizer" data-topic="DBOperations" data-id="get_table"><a href='../../SBPoptimizer/html/DBOperations.html#method-DBOperations-get_table'><code>SBPoptimizer::DBOperations$get_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SBPoptimizer" data-topic="DBOperations" data-id="initialize"><a href='../../SBPoptimizer/html/DBOperations.html#method-DBOperations-initialize'><code>SBPoptimizer::DBOperations$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SBPoptimizer" data-topic="DBOperations" data-id="insert_record"><a href='../../SBPoptimizer/html/DBOperations.html#method-DBOperations-insert_record'><code>SBPoptimizer::DBOperations$insert_record()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SBPoptimizer" data-topic="DBOperations" data-id="update_record"><a href='../../SBPoptimizer/html/DBOperations.html#method-DBOperations-update_record'><code>SBPoptimizer::DBOperations$update_record()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceOperations-update_compliance_score"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceOperations-update_compliance_score}{}}}
\subsection{Method \code{update_compliance_score()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceOperations$update_compliance_score(
  business_id,
  score,
  model_version,
  features_json = NULL,
  cluster_label = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{business_id}}{Business ID}

\item{\code{score}}{Compliance score (0-1)}

\item{\code{model_version}}{Model version string}

\item{\code{features_json}}{Features used in scoring (optional)}

\item{\code{cluster_label}}{Risk cluster label (optional)
Get compliance scores with filtering}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceOperations-get_compliance_scores"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceOperations-get_compliance_scores}{}}}
\subsection{Method \code{get_compliance_scores()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceOperations$get_compliance_scores(
  min_score = NULL,
  max_score = NULL,
  cluster_label = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{min_score}}{Minimum score filter (optional)}

\item{\code{max_score}}{Maximum score filter (optional)}

\item{\code{cluster_label}}{Cluster label filter (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame of compliance scores
Get high-risk businesses
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceOperations-get_high_risk_businesses"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceOperations-get_high_risk_businesses}{}}}
\subsection{Method \code{get_high_risk_businesses()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceOperations$get_high_risk_businesses(threshold = 0.3)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{threshold}}{Risk threshold (default 0.3)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame of high-risk businesses
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComplianceOperations-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ComplianceOperations-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComplianceOperations$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
