% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{create_db_pool}
\alias{create_db_pool}
\title{Create Database Connection Pool}
\usage{
create_db_pool(config_name = Sys.getenv("GOLEM_CONFIG_ACTIVE", "default"))
}
\arguments{
\item{config_name}{Configuration environment name}
}
\value{
Database connection pool object
}
\description{
Create Database Connection Pool
}
