% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_operations.R
\name{DBOperations}
\alias{DBOperations}
\title{Base Database Operations Class}
\description{
R6 class providing basic database operations
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pool}}{Database connection pool
Initialize the database operations class}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DBOperations-new}{\code{DBOperations$new()}}
\item \href{#method-DBOperations-get_table}{\code{DBOperations$get_table()}}
\item \href{#method-DBOperations-insert_record}{\code{DBOperations$insert_record()}}
\item \href{#method-DBOperations-update_record}{\code{DBOperations$update_record()}}
\item \href{#method-DBOperations-delete_record}{\code{DBOperations$delete_record()}}
\item \href{#method-DBOperations-clone}{\code{DBOperations$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBOperations-new"></a>}}
\if{latex}{\out{\hypertarget{method-DBOperations-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBOperations$new(pool)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pool}}{Database connection pool object
Get a database table as a dplyr tbl object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBOperations-get_table"></a>}}
\if{latex}{\out{\hypertarget{method-DBOperations-get_table}{}}}
\subsection{Method \code{get_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBOperations$get_table(table_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table_name}}{Character string of table name}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A dplyr tbl object
Insert a record into a table
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBOperations-insert_record"></a>}}
\if{latex}{\out{\hypertarget{method-DBOperations-insert_record}{}}}
\subsection{Method \code{insert_record()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBOperations$insert_record(table_name, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table_name}}{Character string of table name}

\item{\code{data}}{Data frame with record data
Update a record in a table}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBOperations-update_record"></a>}}
\if{latex}{\out{\hypertarget{method-DBOperations-update_record}{}}}
\subsection{Method \code{update_record()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBOperations$update_record(
  table_name,
  id,
  data,
  id_column = paste0(table_name, "_id")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table_name}}{Character string of table name}

\item{\code{id}}{Record ID to update}

\item{\code{data}}{Named list of columns and values to update}

\item{\code{id_column}}{Character string of ID column name
Delete a record from a table}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBOperations-delete_record"></a>}}
\if{latex}{\out{\hypertarget{method-DBOperations-delete_record}{}}}
\subsection{Method \code{delete_record()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBOperations$delete_record(
  table_name,
  id,
  id_column = paste0(table_name, "_id")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table_name}}{Character string of table name}

\item{\code{id}}{Record ID to delete}

\item{\code{id_column}}{Character string of ID column name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DBOperations-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DBOperations-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBOperations$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
