% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{execute_sql_file}
\alias{execute_sql_file}
\title{Execute SQL File with Better Parsing}
\usage{
execute_sql_file(pool, sql_file)
}
\arguments{
\item{pool}{Database connection pool}

\item{sql_file}{Path to SQL file}
}
\description{
Execute SQL File with Better Parsing
}
