% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{get_db_config}
\alias{get_db_config}
\title{Get Database Configuration}
\usage{
get_db_config(config_name = Sys.getenv("GOLEM_CONFIG_ACTIVE", "default"))
}
\arguments{
\item{config_name}{Configuration environment name}
}
\value{
List with database configuration
}
\description{
Get Database Configuration
}
