% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{initialize_db}
\alias{initialize_db}
\title{Initialize Database Connection}
\usage{
initialize_db(config_name = Sys.getenv("GOLEM_CONFIG_ACTIVE", "default"))
}
\arguments{
\item{config_name}{Configuration environment name}
}
\value{
Database connection pool
}
\description{
Initialize Database Connection
}
