% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_operations.R
\name{InspectionOperations}
\alias{InspectionOperations}
\title{Inspection Operations Class}
\description{
R6 class for inspection management operations
}
\section{Super class}{
\code{\link[SBPoptimizer:DBOperations]{SBPoptimizer::DBOperations}} -> \code{InspectionOperations}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InspectionOperations-create_inspection}{\code{InspectionOperations$create_inspection()}}
\item \href{#method-InspectionOperations-get_inspections_by_business}{\code{InspectionOperations$get_inspections_by_business()}}
\item \href{#method-InspectionOperations-get_inspections_by_inspector}{\code{InspectionOperations$get_inspections_by_inspector()}}
\item \href{#method-InspectionOperations-get_inspection_statistics}{\code{InspectionOperations$get_inspection_statistics()}}
\item \href{#method-InspectionOperations-clone}{\code{InspectionOperations$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SBPoptimizer" data-topic="DBOperations" data-id="delete_record"><a href='../../SBPoptimizer/html/DBOperations.html#method-DBOperations-delete_record'><code>SBPoptimizer::DBOperations$delete_record()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SBPoptimizer" data-topic="DBOperations" data-id="get_table"><a href='../../SBPoptimizer/html/DBOperations.html#method-DBOperations-get_table'><code>SBPoptimizer::DBOperations$get_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SBPoptimizer" data-topic="DBOperations" data-id="initialize"><a href='../../SBPoptimizer/html/DBOperations.html#method-DBOperations-initialize'><code>SBPoptimizer::DBOperations$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SBPoptimizer" data-topic="DBOperations" data-id="insert_record"><a href='../../SBPoptimizer/html/DBOperations.html#method-DBOperations-insert_record'><code>SBPoptimizer::DBOperations$insert_record()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SBPoptimizer" data-topic="DBOperations" data-id="update_record"><a href='../../SBPoptimizer/html/DBOperations.html#method-DBOperations-update_record'><code>SBPoptimizer::DBOperations$update_record()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InspectionOperations-create_inspection"></a>}}
\if{latex}{\out{\hypertarget{method-InspectionOperations-create_inspection}{}}}
\subsection{Method \code{create_inspection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InspectionOperations$create_inspection(
  business_id,
  inspector_id,
  notes = NULL,
  status,
  photo_url = NULL,
  latitude = NULL,
  longitude = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{business_id}}{Business ID}

\item{\code{inspector_id}}{Inspector user ID}

\item{\code{notes}}{Inspection notes (optional)}

\item{\code{status}}{Inspection status}

\item{\code{photo_url}}{Photo URL (optional)}

\item{\code{latitude}}{GPS latitude (optional)}

\item{\code{longitude}}{GPS longitude (optional)
Get inspections by business}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InspectionOperations-get_inspections_by_business"></a>}}
\if{latex}{\out{\hypertarget{method-InspectionOperations-get_inspections_by_business}{}}}
\subsection{Method \code{get_inspections_by_business()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InspectionOperations$get_inspections_by_business(business_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{business_id}}{Business ID}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame of inspections
Get inspections by inspector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InspectionOperations-get_inspections_by_inspector"></a>}}
\if{latex}{\out{\hypertarget{method-InspectionOperations-get_inspections_by_inspector}{}}}
\subsection{Method \code{get_inspections_by_inspector()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InspectionOperations$get_inspections_by_inspector(
  inspector_id,
  start_date = NULL,
  end_date = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inspector_id}}{Inspector user ID}

\item{\code{start_date}}{Start date filter (optional)}

\item{\code{end_date}}{End date filter (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame of inspections
Get inspection statistics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InspectionOperations-get_inspection_statistics"></a>}}
\if{latex}{\out{\hypertarget{method-InspectionOperations-get_inspection_statistics}{}}}
\subsection{Method \code{get_inspection_statistics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InspectionOperations$get_inspection_statistics()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Data frame with inspection statistics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InspectionOperations-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InspectionOperations-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InspectionOperations$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
