% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_database_admin.R
\name{mod_database_admin_ui}
\alias{mod_database_admin_ui}
\alias{mod_database_admin_server}
\title{Database Administration Module}
\usage{
mod_database_admin_ui(id)

mod_database_admin_server(id, db_pool)
}
\arguments{
\item{id}{character Module ID}
}
\description{
A Shiny module for database administration tasks
}
\keyword{internal}
