% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{parse_sql_statements}
\alias{parse_sql_statements}
\title{Parse SQL statements respecting dollar-quoted strings}
\usage{
parse_sql_statements(sql_content)
}
\arguments{
\item{sql_content}{Full SQL content as string}
}
\value{
Vector of SQL statements
}
\description{
Parse SQL statements respecting dollar-quoted strings
}
\keyword{internal}
