% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_operations.R
\name{PermitOperations}
\alias{PermitOperations}
\title{Permit Operations Class}
\description{
R6 class for permit management operations
}
\section{Super class}{
\code{\link[SBPoptimizer:DBOperations]{SBPoptimizer::DBOperations}} -> \code{PermitOperations}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PermitOperations-create_permit}{\code{PermitOperations$create_permit()}}
\item \href{#method-PermitOperations-get_permits_by_business}{\code{PermitOperations$get_permits_by_business()}}
\item \href{#method-PermitOperations-get_expiring_permits}{\code{PermitOperations$get_expiring_permits()}}
\item \href{#method-PermitOperations-calculate_permit_fee}{\code{PermitOperations$calculate_permit_fee()}}
\item \href{#method-PermitOperations-update_permit_status}{\code{PermitOperations$update_permit_status()}}
\item \href{#method-PermitOperations-clone}{\code{PermitOperations$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SBPoptimizer" data-topic="DBOperations" data-id="delete_record"><a href='../../SBPoptimizer/html/DBOperations.html#method-DBOperations-delete_record'><code>SBPoptimizer::DBOperations$delete_record()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SBPoptimizer" data-topic="DBOperations" data-id="get_table"><a href='../../SBPoptimizer/html/DBOperations.html#method-DBOperations-get_table'><code>SBPoptimizer::DBOperations$get_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SBPoptimizer" data-topic="DBOperations" data-id="initialize"><a href='../../SBPoptimizer/html/DBOperations.html#method-DBOperations-initialize'><code>SBPoptimizer::DBOperations$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SBPoptimizer" data-topic="DBOperations" data-id="insert_record"><a href='../../SBPoptimizer/html/DBOperations.html#method-DBOperations-insert_record'><code>SBPoptimizer::DBOperations$insert_record()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SBPoptimizer" data-topic="DBOperations" data-id="update_record"><a href='../../SBPoptimizer/html/DBOperations.html#method-DBOperations-update_record'><code>SBPoptimizer::DBOperations$update_record()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PermitOperations-create_permit"></a>}}
\if{latex}{\out{\hypertarget{method-PermitOperations-create_permit}{}}}
\subsection{Method \code{create_permit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PermitOperations$create_permit(
  business_id,
  permit_year,
  fee_amount,
  issue_date = Sys.Date(),
  expiry_date = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{business_id}}{Business ID}

\item{\code{permit_year}}{Permit year}

\item{\code{fee_amount}}{Fee amount}

\item{\code{issue_date}}{Issue date (defaults to today)}

\item{\code{expiry_date}}{Expiry date (defaults to end of permit year)
Get permits by business}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PermitOperations-get_permits_by_business"></a>}}
\if{latex}{\out{\hypertarget{method-PermitOperations-get_permits_by_business}{}}}
\subsection{Method \code{get_permits_by_business()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PermitOperations$get_permits_by_business(business_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{business_id}}{Business ID}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame of permits
Get expiring permits
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PermitOperations-get_expiring_permits"></a>}}
\if{latex}{\out{\hypertarget{method-PermitOperations-get_expiring_permits}{}}}
\subsection{Method \code{get_expiring_permits()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PermitOperations$get_expiring_permits(days_ahead = 30)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{days_ahead}}{Number of days ahead to check (default 30)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame of expiring permits
Calculate permit fee based on business size and category
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PermitOperations-calculate_permit_fee"></a>}}
\if{latex}{\out{\hypertarget{method-PermitOperations-calculate_permit_fee}{}}}
\subsection{Method \code{calculate_permit_fee()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PermitOperations$calculate_permit_fee(business_size, business_category)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{business_size}}{Business size (Small, Medium, Large)}

\item{\code{business_category}}{Business category}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Numeric fee amount
Update permit status
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PermitOperations-update_permit_status"></a>}}
\if{latex}{\out{\hypertarget{method-PermitOperations-update_permit_status}{}}}
\subsection{Method \code{update_permit_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PermitOperations$update_permit_status(permit_id, status, penalty_amount = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{permit_id}}{Permit ID}

\item{\code{status}}{New status}

\item{\code{penalty_amount}}{Penalty amount (default 0)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PermitOperations-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PermitOperations-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PermitOperations$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
