% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_operations.R
\name{UserOperations}
\alias{UserOperations}
\title{User Operations Class}
\description{
R6 class for user management operations
}
\section{Super class}{
\code{\link[SBPoptimizer:DBOperations]{SBPoptimizer::DBOperations}} -> \code{UserOperations}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserOperations-create_user}{\code{UserOperations$create_user()}}
\item \href{#method-UserOperations-get_user_by_email}{\code{UserOperations$get_user_by_email()}}
\item \href{#method-UserOperations-get_user_by_id}{\code{UserOperations$get_user_by_id()}}
\item \href{#method-UserOperations-authenticate_user}{\code{UserOperations$authenticate_user()}}
\item \href{#method-UserOperations-update_user_activity}{\code{UserOperations$update_user_activity()}}
\item \href{#method-UserOperations-clone}{\code{UserOperations$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SBPoptimizer" data-topic="DBOperations" data-id="delete_record"><a href='../../SBPoptimizer/html/DBOperations.html#method-DBOperations-delete_record'><code>SBPoptimizer::DBOperations$delete_record()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SBPoptimizer" data-topic="DBOperations" data-id="get_table"><a href='../../SBPoptimizer/html/DBOperations.html#method-DBOperations-get_table'><code>SBPoptimizer::DBOperations$get_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SBPoptimizer" data-topic="DBOperations" data-id="initialize"><a href='../../SBPoptimizer/html/DBOperations.html#method-DBOperations-initialize'><code>SBPoptimizer::DBOperations$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SBPoptimizer" data-topic="DBOperations" data-id="insert_record"><a href='../../SBPoptimizer/html/DBOperations.html#method-DBOperations-insert_record'><code>SBPoptimizer::DBOperations$insert_record()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SBPoptimizer" data-topic="DBOperations" data-id="update_record"><a href='../../SBPoptimizer/html/DBOperations.html#method-DBOperations-update_record'><code>SBPoptimizer::DBOperations$update_record()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserOperations-create_user"></a>}}
\if{latex}{\out{\hypertarget{method-UserOperations-create_user}{}}}
\subsection{Method \code{create_user()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserOperations$create_user(
  name,
  email,
  phone = NULL,
  password_hash,
  role = "officer"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{User's full name}

\item{\code{email}}{User's email address}

\item{\code{phone}}{User's phone number (optional)}

\item{\code{password_hash}}{Hashed password}

\item{\code{role}}{User role (admin, officer, inspector, finance)
Get user by email}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserOperations-get_user_by_email"></a>}}
\if{latex}{\out{\hypertarget{method-UserOperations-get_user_by_email}{}}}
\subsection{Method \code{get_user_by_email()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserOperations$get_user_by_email(email)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email}}{User's email address}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame with user information
Get user by ID
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserOperations-get_user_by_id"></a>}}
\if{latex}{\out{\hypertarget{method-UserOperations-get_user_by_id}{}}}
\subsection{Method \code{get_user_by_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserOperations$get_user_by_id(user_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user_id}}{User ID}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame with user information
Authenticate user
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserOperations-authenticate_user"></a>}}
\if{latex}{\out{\hypertarget{method-UserOperations-authenticate_user}{}}}
\subsection{Method \code{authenticate_user()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserOperations$authenticate_user(email, password_hash)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{email}}{User's email}

\item{\code{password_hash}}{Hashed password to verify}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
User data frame if authenticated, NULL otherwise
Update user last activity
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserOperations-update_user_activity"></a>}}
\if{latex}{\out{\hypertarget{method-UserOperations-update_user_activity}{}}}
\subsection{Method \code{update_user_activity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserOperations$update_user_activity(user_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user_id}}{User ID}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserOperations-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserOperations-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserOperations$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
